requirements.txt
altair==5.5.0
annotated-types==0.7.0
anywidget==0.9.13
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
branca==0.8.1
cachetools==5.5.0
captcha==0.6.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dateutils==0.6.12
decorator==5.1.1
duckdb==1.1.3
et_xmlfile==2.0.0
executing==2.2.0
extra-streamlit-components==0.1.71
folium==0.19.4
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
gw_dsl_parser==0.1.49.1
idna==3.10
importlib_resources==6.5.2
ipylab==1.0.0
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kanaries_track==0.0.5
kiwisolver==1.4.8
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
narwhals==1.23.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
psygnal==0.11.1
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pygwalker==0.4.9.13
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
quickjs==1.19.4
referencing==0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
segment-analytics-python==2.2.3
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
sqlglot==26.3.8
stack-data==0.6.3
streamlit==1.41.1
streamlit-authenticator==0.4.1
streamlit_folium==0.24.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wasmtime==29.0.0
watchdog==6.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
win32_setctime==1.2.0
xyzservices==2025.1.0
